Created separate auto command and routine packages
